Rename movie package to group
